added new seeders and replaced movie code with slug

+++ seeders/customer.seeder.js
@@ -3,13 +3,13 @@ const { User } = require("../models");
 
 module.exports = {
     seed: async (count = 1) => {
-        await User.deleteMany();
+        await User.deleteMany({ role: 'customer' });
         for(let i = 0; i < count; i++) {
             await User.create({
                 name: faker.name.fullName(),
                 email: faker.internet.email(),
                 password: '123456',
-                role: 'admin'
+                role: 'customer'
             }).then(data => {
                 console.log(`Seeded: ${data}`)
             }).catch(err => console.log(`Error seeding: ${err}`));
